Add watches for ArgoCD and Rancher to Keycloak to set up SSO clients
