+++ platform-operator/controllers/verrazzano/component/keycloak/module_integration.go
@@ -5,15 +5,64 @@ package keycloak
 
 import (
 	"github.com/verrazzano/verrazzano-modules/pkg/controller/spi/controllerspi"
+	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	cmconstants "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/constants"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common/watch"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/fluentoperator"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/istio"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/mysql"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/nginx"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 )
+
+// valuesConfig Structure for the translated effective Verrazzano CR values to Module CR Helm values
+type valuesConfig struct {
+	Ingress         *vzapi.IngressNginxComponent `json:"ingress,omitempty"`
+	DNS             *vzapi.DNSComponent          `json:"dns,omitempty"`
+	EnvironmentName string                       `json:"environmentName,omitempty"`
+}
 
 // GetWatchDescriptors returns the list of WatchDescriptors for objects being watched by the component
 func (c KeycloakComponent) GetWatchDescriptors() []controllerspi.WatchDescriptor {
-	return watch.GetModuleInstalledWatches([]string{istio.ComponentName, nginx.ComponentName, cmconstants.CertManagerComponentName, mysql.ComponentName, fluentoperator.ComponentName})
+	return watch.GetModuleInstalledWatches([]string{
+		istio.ComponentName,
+		nginx.ComponentName,
+		cmconstants.CertManagerComponentName,
+		mysql.ComponentName,
+		fluentoperator.ComponentName,
+		// ArgoCD and Rancher require Keycloak to be re-reconciled to build the client IDs
+		common.ArgoCDName,
+		common.RancherName,
+	})
+}
+
+// GetModuleConfigAsHelmValues returns an unstructured JSON valuesConfig representing the portion of the Verrazzano CR that corresponds to the module
+func (c KeycloakComponent) GetModuleConfigAsHelmValues(effectiveCR *vzapi.Verrazzano) (*apiextensionsv1.JSON, error) {
+	if effectiveCR == nil {
+		return nil, nil
+	}
+
+	configSnippet := valuesConfig{
+		EnvironmentName: effectiveCR.Spec.EnvironmentName,
+	}
+
+	dns := effectiveCR.Spec.Components.DNS
+	if dns != nil {
+		configSnippet.DNS = &vzapi.DNSComponent{
+			External:         dns.External,
+			InstallOverrides: vzapi.InstallOverrides{}, // always ignore the overrides here, those are handled separately
+			OCI:              dns.OCI,
+			Wildcard:         dns.Wildcard,
+		}
+	}
+
+	nginx := effectiveCR.Spec.Components.Ingress
+	if nginx != nil {
+		configSnippet.Ingress = nginx.DeepCopy()
+		configSnippet.Ingress.InstallOverrides.ValueOverrides = []vzapi.Overrides{}
+	}
+
+	return spi.NewModuleConfigHelmValuesWrapper(configSnippet)
 }
