--- platform-operator/controllers/verrazzano/component/keycloak/modules_integration.go
@@ -1,46 +1,0 @@
-// Copyright (c) 2023, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-
-package keycloak
-
-import (
-	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-)
-
-// valuesConfig Structure for the translated effective Verrazzano CR values to Module CR Helm values
-type valuesConfig struct {
-	Ingress         *vzapi.IngressNginxComponent `json:"ingress,omitempty"`
-	DNS             *vzapi.DNSComponent          `json:"dns,omitempty"`
-	EnvironmentName string                       `json:"environmentName,omitempty"`
-}
-
-// GetModuleConfigAsHelmValues returns an unstructured JSON valuesConfig representing the portion of the Verrazzano CR that corresponds to the module
-func (c KeycloakComponent) GetModuleConfigAsHelmValues(effectiveCR *vzapi.Verrazzano) (*apiextensionsv1.JSON, error) {
-	if effectiveCR == nil {
-		return nil, nil
-	}
-
-	configSnippet := valuesConfig{
-		EnvironmentName: effectiveCR.Spec.EnvironmentName,
-	}
-
-	dns := effectiveCR.Spec.Components.DNS
-	if dns != nil {
-		configSnippet.DNS = &vzapi.DNSComponent{
-			External:         dns.External,
-			InstallOverrides: vzapi.InstallOverrides{}, // always ignore the overrides here, those are handled separately
-			OCI:              dns.OCI,
-			Wildcard:         dns.Wildcard,
-		}
-	}
-
-	nginx := effectiveCR.Spec.Components.Ingress
-	if nginx != nil {
-		configSnippet.Ingress = nginx.DeepCopy()
-		configSnippet.Ingress.InstallOverrides.ValueOverrides = []vzapi.Overrides{}
-	}
-
-	return spi.NewModuleConfigHelmValuesWrapper(configSnippet)
-}
